// src/modules/flood-damages/enums/damage-category.enum.ts
export enum DamageCategory {
  ECONOMIC = 'ECONOMIC', // Thiệt hại kinh tế (cây trồng, vật nuôi…)
  PROPERTY = 'PROPERTY', // Thiệt hại tài sản (nhà cửa, đồ đạc…)
  BUSINESS = 'BUSINESS', // Ảnh hưởng hoạt động kinh doanh
  HEALTH = 'HEALTH', // Ảnh hưởng sức khỏe
  FATALITY = 'FATALITY', // Thương vong (tử vong)
  OTHER = 'OTHER', // Khác
}