// src/modules/verifications/enums/verification.enum.ts
export enum VerificationStatus {
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
  COMPLETED = 'COMPLETED',
}

export enum VerificationType {
  RESIDENT_REGISTRATION = 'RESIDENT_REGISTRATION',
  VEHICLE_REGISTRATION = 'VEHICLE_REGISTRATION',
  TEMPORARY_RESIDENT = 'TEMPORARY_RESIDENT',
  REFLECTION = 'REFLECTION',
  OTHER = 'OTHER',
}